requirements.txt
streamlit
langchain-core
langchain-ollama
langchain-community
python-dotenv